[Jetsurvey] Changed to use Navigation Compose
This eliminates all fragments in favor of navigating at the composable
level. I also updated the screen-level composables to take in explicit
lambdas rather than event handlers to follow best practices.

One notable change here is that there is no longer a fragment from which
to launch the camera. Instead, the PhotoQuestion uses
rememberLauncherForActivityResult. This also uses TakePicturePreview,
which doesn't require a pre-registered URI, so the file permission is no
longer necessary (the contract returns a Bitmap directly).